pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "auth0-fastapi>=1.0.0b4",
    "auth0-python>=4.10.0",
    "authlib>=1.6.3",
    "cryptography>=45.0.6",
    "dotenv>=0.9.9",
    "fastapi[all]>=0.115.14",
    "m2r2>=0.3.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic>=2.11.7",
    "pyjwt==2.10.1",
    "selenium-driverless>=1.9.4",
    "setuptools==80.9.0",
    "sphinx>=8.2.3",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "undetected-chromedriver>=3.5.5",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
]

[tool.uv.workspace]
members = [
    ".",
    ".",
]

[tool.uv.sources]
backend = { workspace = true }

[dependency-groups]
dev = [
    "backend",
]
